Route all calculations through a single endpoint
All calculations are directed to the /calculate endpoint. Every
request object should contain a 'wheel' object and may contain
multiple calculation request objects such as 'tension' or
'stiffness'. The result will contain objects with the same name as
the request objects, each with a boolean 'success' flag and
optional 'error' description. As long as the wheel object is present
and the request object is well-formed, the result code will be
200, although errors may have occurred on individual calculations.